requirements.txt
streamlit~=1.42.2
numpy~=2.2.3
pillow~=11.1.0
opencv-python
